@@ -43,8 +43,8 @@
     "@types/lodash": "4.14.182",
     "@types/luxon": "2.3.2",
     "@types/prismjs": "1.26.0",
-    "@types/react": "17.0.47",
-    "@types/react-dom": "17.0.17",
+    "@types/react": "18.0.5",
+    "@types/react-dom": "18.0.5",
     "@types/react-router-dom": "5.3.3",
     "@types/react-window": "1.8.5",
     "@types/sanitize-html": "2.6.2",
@@ -98,8 +98,8 @@
     "lodash": "4.17.21",
     "luxon": "2.4.0",
     "prismjs": "1.28.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "react-router-dom": "6.3.0",
     "react-window": "1.8.7",
     "sanitize-html": "2.7.0",
